Add function to find a stop
This will add a function pyogt.find_stop(name) that will return an
OGTStop object using OGTRest (rest api). A test was also added that
find known stops using their name and check that it exist. This test
could potentially break it the future if any of the currently known
stops would change name or be removed.